190228 15:27 form validation

+++ src/main/java/com/douzone/mysite/controller/GuestbookController.java
@@ -1,12 +1,15 @@
 package com.douzone.mysite.controller;
 
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
 
 import com.douzone.mysite.service.GuestbookService;
 import com.douzone.mysite.vo.GuestbookVo;
@@ -30,12 +33,23 @@ public class GuestbookController {
 		return "redirect:/guestbook/list";
 	}
 	
-	@RequestMapping(value="/delete", method=RequestMethod.GET)
-	public String delete(@RequestParam(value="no", required=true) Long no) {
+	@RequestMapping(value="/delete/{no}", method=RequestMethod.GET)
+	public String delete(@PathVariable("no") Long no,
+			@ModelAttribute GuestbookVo guestbookVo,
+			Model model) {
+		model.addAttribute("no", no);
 		return "guestbook/deleteform";
 	}
 	@RequestMapping(value="/delete", method=RequestMethod.POST)
-	public String delete(@ModelAttribute GuestbookVo guestbookVo) {
+	public String delete(
+			@ModelAttribute @Valid GuestbookVo guestbookVo,
+			BindingResult result,
+			Model model) {
+		if(result.hasErrors()) {
+			model.addAttribute(result.getModel());
+			return "redirect:/guestbook/delete/"+guestbookVo.getNo();
+		}
+		
 		guestbookService.delete(guestbookVo);
 		return "redirect:/guestbook/list";
 	}
